swap out fsnotify for inotify
fsnotify has a bug where it generates an lstat call for every inotify
event.  You also can't select the inotify mask.  So if you watch a
directory with fsnotify, you inadvertently generate an lstat syscall for
every modification made to every file in that directory.  That's not
something we want.

Fortunately Rob Pike had written the original inotify library that the
fsnotify libraryis based off of, so I'm just using the Rob Pike
implementation, which is much better.

This means that Lumberjack only supports Linux properly at this point in
time.