buffer publisher channel
when a publisher fails to send a page, it puts the page back on the
channel of pending pages.  if that send blocks, the publisher doesn't
get to attempt to reconnect.  if *all* logstash servers are down, N
publishers are guaranteed to be able to retry their connections.  if N
is 1, like it used to be, then the cluster wouldn't bounce back until a
*specific* node went back online.  setting it to the number of
publishers means that we'll never deadlock our publishers in an outage.